[RecursiveDescent] Eliminate indirection in one-arm cases
This modifies Mix implementation to recognize when an Alternative has
one symbol, or Symbol consists of one alternative, and eliminates the
intermediate function entirely if that is the case. Note that this does
not handle rules like `A -> B` when a symbol has one alternative
consisting of some other symbol.